vision_model/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
scikit-learn
numpy 